Backport ZOOKEEPER-1045.
Reason: Security
Author: Michael Han
Ref: CDH-41446

This commit back ports ZOOKEEPER-1045 to CDH. Besides ZOOKEEPER-1045, this commit also contains a minimum set of code of ZOOKEEPER-1506 that ZOOKEEPER-1045 depends on. The reason we need ZOOKEEPER-1506 is ZOOKEEPER-1045 needs to persist the hostname parameter and previously (before ZOOKEEPER-1506, this parameter is simply converted to an InetSocketAddress during parameter parsing step, so we lost it permanently on ZK server side after server starts up..) this parameter was not persisted. ZK Server needs this parameter (which is essentially the server name) to do authorization check. The code involved in ZOOKEEPER-1506 is updated in such a way that no semantic change is involved: unlike ZOOKEEPER-1506 which will always try resolving server name by issuing a DNS lookup, we keep the old behavior by creating the InetSocketAddress directly from the configured parameter (the hostname, port) from zoo.cfg.

* ZOOKEEPER-1045: https://github.com/apache/zookeeper/commit/8a06bd1ccef382461c7b0a63f2012f4aeac90753

Change-Id: I5ff587cc1776ca3b4c64dd5105d3fe8aad7738da