CDH-44828: The documentation should have the reference to ZooKeeper 3.5 clarified or removed
When ZOOKEEPER-1161 was backported, the PDF and HTML documentation was not recompiled from source.
The documentation was recompiled when CDH-44819 was committed, which made the newly added section describing
ZOOKEEPER-1161 visible only since then. This has a reference to ZooKeeper 3.5 which is incorrect as
ZooKeeper in CDH is currently based on ZooKeeper 3.4.
The purpose of this change is to clarify that the change which the documentation describes is backported from
ZooKeeper 3.5 and not the version of ZooKeeper changed to 3.5.

Change-Id: I152545b5a84198527c261b2205d2da52e86aaa2f